# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: AgentDash API
  version: v1
  description: |-
    The public HTTP API contract. Every operation here is a promise: it is not removed or renamed without a deprecation cycle.

    Routes not in this document are internal and may change without notice; the route index lists them.

    Generated by scripts/docs/generate-openapi.mjs from docs/api/contract.json, the zod validators and the TypeScript types in @paperclipai/shared. Do not edit by hand.
servers:
  - url: "{instanceUrl}"
    description: Your AgentDash instance.
    variables:
      instanceUrl:
        default: https://your-instance.example
        description: Your AgentDash instance's address
security:
  - bearerBoardKey: []
  - bearerAgentKey: []
  - session: []
...
paths:
...
  /api/companies/{companyId}/activity:
    get:
      operationId: listActivity
      summary: List a company's activity log, filtered by agent, entity or time.
      tags:
        - activity
      parameters:
        - name: companyId
          in: path
          required: true
          description: Company id (UUID).
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Rows to return. Default 100, clamped to 1–500.
          schema:
            type: integer
        - name: since
          in: query
          required: false
          description: Only rows at or after this ISO 8601 timestamp; an unparseable value is a 400.
          schema:
            type: string
            format: date-time
        - name: agentId
          in: query
          required: false
          description: Only rows about this agent.
          schema:
            type: string
        - name: entityType
          in: query
          required: false
          description: Only rows about this kind of entity, such as issue.
          schema:
            type: string
        - name: entityId
          in: query
          required: false
          description: Only rows about this entity.
          schema:
            type: string
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
      security:
        - bearerBoardKey: []
        - bearerAgentKey: []
        - session: []
      x-stability: stable
...
components:
  securitySchemes:
    bearerBoardKey:
      type: http
      scheme: bearer
      description: "A board API key (`pcp_board_…`): acts as the person who approved it, with that person's company memberships and nothing narrower — board keys have no scopes. Expires 30 days after it is minted. Minted by the CLI sign-in flow; see the API keys page. (server/src/middleware/auth.ts, `board_key`.)"
    bearerAgentKey:
      type: http
      scheme: bearer
      description: "An agent API key (`pcp_…`): acts as one agent in one company. Also accepted as the `x-agent-key` header when no Authorization header is sent. A run-scoped agent JWT issued to a local adapter is accepted the same way. (server/src/middleware/auth.ts, `agent_key` and `agent_jwt`.)"
    session:
      type: apiKey
      in: cookie
      name: paperclip-default.session_token
      description: The web app's signed-in session cookie. The name is per instance (`<prefix>.session_token`, `__Secure-` prefixed over https); browsers send it automatically. Meant for the web app, not for integrations — use a board key. (server/src/middleware/auth.ts, `session`.)
    assistantOAuth:
      type: oauth2
      description: "OAuth 2.1 with PKCE, for an assistant acting for one person in one company (`pcpa_…` access tokens). Discovery: `/.well-known/oauth-protected-resource` and `/.well-known/oauth-authorization-server`. A grant reaches only the assistant routes its scopes allow; anything else is 403. (server/src/middleware/auth.ts, `assistant_grant`.)"
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          refreshUrl: /oauth/token
          scopes:
            agentdash:read: Read the company's work.
            agentdash:work: Create and move work, wake agents.
            agentdash:decide: Resolve approvals and hire. Opt-in at consent; never granted by default.
    bearerBridgeEndpoint:
      type: http
      scheme: bearer
      description: An enrolled machine's endpoint token. Resolved only on the bridge's own routes; on any other path it is treated as no credential at all. (server/src/middleware/auth.ts, `bridge_endpoint`.)
  responses:
    BadRequest:
      description: The request is malformed. `Invalid identifier` when a path or query id is not a UUID (server/src/middleware/error-handler.ts); `Validation error` with `details` when the body fails its schema.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: No credential, or the credential did not resolve to an actor.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: "The caller is authenticated but may not do this: a missing permission, a read-only principal, or a credential that cannot reach this route (an assistant grant without the required scope answers `insufficient_scope`)."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: "No such resource, **or one the caller is not allowed to see**. Visibility answers 404, never 403: a 403 on a guessed id would confirm the thing exists (server/src/routes/visibility.ts)."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RateLimited:
      description: "Rate limited (server/src/middleware/rate-limit.ts). Under /api, state-changing requests are limited per actor (or per IP when unauthenticated): 200 per 15 minutes by default, lower on auth, billing, invite and trial routes. Authenticated reads are not counted. The response carries `Retry-After` and the draft-7 `RateLimit` headers. Instances may configure other limits."
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RateLimitError"
  schemas:
    ActivityEvent:
      type: object
      properties:
        id:
          type: string
        companyId:
          type: string
        actorType:
          type: string
          enum:
            - agent
            - user
            - system
            - plugin
        actorId:
          type: string
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: string
        agentId:
          type:
            - string
            - "null"
        runId:
          type:
            - string
            - "null"
        details:
          anyOf:
            - type: object
              additionalProperties: {}
            - type: "null"
        origin:
          type:
            - string
            - "null"
          enum:
            - server
            - manual
            - null
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - companyId
        - actorType
        - actorId
        - action
        - entityType
        - entityId
        - agentId
        - runId
        - details
        - createdAt
...
    Error:
      type: object
      description: Every error is JSON with a human-readable `error`. Some carry `details` (a validation error lists the failing fields) or a machine-readable `code`.
      properties:
        error:
          type: string
        details: {}
        code:
          type: string
      required:
        - error
...
    RateLimitError:
      type: object
      properties:
        error:
          type: string
          const: Rate limited
        retryAfter:
          type: integer
          description: Seconds until the window resets (900).
      required:
        - error
        - retryAfter
...